requirements.txt
albumentations==1.3.0
asttokens==2.2.1
backcall==0.2.0
bzip2==1.0.8
ca-certificates==2023.01.10
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
filelock==3.9.0
idna==3.4
imageio==2.30.0
ipykernel==6.23.1
ipython==8.13.2
jedi==0.18.2
jinja2==3.1.2
joblib==1.2.0
jupyter-client==8.2.0
jupyter-core==5.3.0
lazy-loader==0.2
libffi==3.4.4
markupsafe==2.1.2
matplotlib-inline==0.1.6
mpmath==1.2.1
nest-asyncio==1.5.6
networkx==3.0
numpy==1.24.1
opencv-python-headless==4.7.0.72
openssl==1.1.1t
packaging==23.1
pandas==2.0.2
parso==0.8.3
pickleshare==0.7.5
pillow==9.3.0
platformdirs==3.5.1
prompt-toolkit==3.0.38
psutil==5.9.5
pure-eval==0.2.2
pygments==2.15.1
python-dateutil==2.8.2
pytz==2023.3
pywavelets==1.4.1
pywin32==306
pyyaml==6.0
pyzmq==25.1.0
qudida==0.0.4
requests==2.28.1
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.10.1
setuptools==67.8.0
six==1.16.0
sqlite==3.41.2
stack-data==0.6.2
sympy==1.11.1
threadpoolctl==3.1.0
tifffile==2023.4.12
tk==8.6.12
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing-extensions==4.4.0
tzdata==2023.3
urllib3==1.26.13
vc==14.2
vs2015_runtime==14.27.29016
wcwidth==0.2.6
wheel==0.38.4
xz==5.4.2
zlib==1.2.13
